Squashed commits for TLS OpenSSL

+++ examples/actors_and_timers/ActorB.cpp
@@ -60,7 +60,7 @@ void ActorB::handleEvent(std::shared_ptr<EventTimerB1> pevt)
 {
     static int s_cabbages =0;
     if (!pevt) return;
-    printf("ActorB::TimerB1 Elapsed - repeats every %ld s\n", (long int) timers.getTimerRepeat<EventTimerB1>().count());
+    printf("ActorB::TimerB1 Elapsed - repeats every %lld s\n", (long long int) timers.getTimerRepeat<EventTimerB1>().count());
     std::shared_ptr<feal::Event> evtCabbage = std::make_shared<EventCabbage>();
     ((EventCabbage*) evtCabbage.get())->set_cabbages(s_cabbages++);
     publishEvent(evtCabbage);
