+++ examples/tcp_echo_clientserver/clienthandler.cpp
@@ -20,12 +20,12 @@ void ClientHandler::setParam(feal::Stream<Server>* p, feal::handle_t fd, char *s
 
 void ClientHandler::initActor(void)
 {
-    printf("ClientHandler(%ld)::initActor\n", (long int) sockfd);
+    printf("ClientHandler(%lld)::initActor\n", (long long int) sockfd);
 }
 
 void ClientHandler::startActor(void)
 {
-    printf("ClientHandler(%ld)::startActor\n", (long int) sockfd);
+    printf("ClientHandler(%lld)::startActor\n", (long long int) sockfd);
     feal::errenum se =feal::FEAL_OK;
     if (stream) se = stream->registerClient<ClientHandler, EvtDataReadAvail, EvtDataWriteAvail, EvtClientShutdown>(this, sockfd);
     if (se != feal::FEAL_OK)
@@ -36,18 +36,18 @@ void ClientHandler::startActor(void)
 
 void ClientHandler::pauseActor(void)
 {
-    printf("ClientHandler(%ld)::pauseActor\n", (long int) sockfd);
+    printf("ClientHandler(%lld)::pauseActor\n", (long long int) sockfd);
 }
 
 void ClientHandler::shutdownActor(void)
 {
-    printf("ClientHandler(%ld)::shutdownActor\n", (long int) sockfd);
+    printf("ClientHandler(%lld)::shutdownActor\n", (long long int) sockfd);
     stream->disconnect_client(sockfd);
 }
 
 void ClientHandler::handleEvent(std::shared_ptr<EvtDataReadAvail> pevt)
 {
-    printf("ClientHandler(%ld)::EvtDataReadAvail\n", (long int) sockfd);
+    printf("ClientHandler(%lld)::EvtDataReadAvail\n", (long long int) sockfd);
     if ((!pevt)||(!stream)) return;
     char buf[30];
     memset(&buf, 0, sizeof(buf));
@@ -56,8 +56,8 @@ void ClientHandler::handleEvent(std::shared_ptr<EvtDataReadAvail> pevt)
     if (se != feal::FEAL_OK) printf("Error receiving: %d\n", se);
     else
     {
-        printf("Received %d bytes \"%s\" from %s\n", bytes, buf, addrstr.c_str());
-        printf("Sending back %d bytes \"%s\" to %s\n", bytes, buf, addrstr.c_str());
+        printf("Received %lld bytes \"%s\" from %s\n", (long long int) bytes, buf, addrstr.c_str());
+        printf("Sending back %lld bytes \"%s\" to %s\n", (long long int) bytes, buf, addrstr.c_str());
         se = stream->send((void*) buf, MIN(strlen(buf) + 1, sizeof(buf)), &bytes, sockfd);
         if (se != feal::FEAL_OK) printf("Error sending \"ClientHandler n\": %d\n", se);
     }
@@ -66,13 +66,13 @@ void ClientHandler::handleEvent(std::shared_ptr<EvtDataReadAvail> pevt)
 void ClientHandler::handleEvent(std::shared_ptr<EvtDataWriteAvail> pevt)
 {
     if (!pevt) return;
-    printf("ClientHandler(%ld)::EvtDataWriteAvail\n", (long int) sockfd);
+    printf("ClientHandler(%lld)::EvtDataWriteAvail\n", (long long int) sockfd);
 }
 
 void ClientHandler::handleEvent(std::shared_ptr<EvtClientShutdown> pevt)
 {
     if (!pevt) return;
-    printf("ClientHandler(%ld)::EvtClientShutdown\n", (long int) sockfd);
+    printf("ClientHandler(%lld)::EvtClientShutdown\n", (long long int) sockfd);
     std::shared_ptr<EvtClientDisconnected> pevt2 = std::make_shared<EvtClientDisconnected>();
     pevt2.get()->fd = sockfd;
     publishEvent(pevt2);
